# Fix AVX2 distance function and build with AVX2 if available on host machine