--- a/src/include/detail/scoring/inner_product_avx.h
+++ b/src/include/detail/scoring/inner_product_avx.h
@@ -155,12 +155,12 @@ template <feature_vector V, feature_vector W>
 inline float avx2_inner_product(const V& a, const W& b) {
   // @todo Align on 256 bit boundaries
   const size_t start = 0;
-  const size_t size_a = a.size();
+  const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
   // Could be uint8_t* or int8_t*
   const auto* a_ptr = a.data();
-  const float* b_ptr = b.data();
+  const auto* b_ptr = b.data();
 
   __m256 vec_sum = _mm256_setzero_ps();
 
@@ -168,8 +168,8 @@ inline float avx2_inner_product(const V& a, const W& b) {
     // Load 8 bytes (either uint8_t or int8_t)
     __m128i vec_a = _mm_loadu_si64((__m64*)(a_ptr + i));
 
-    // Load 8 floats from vector b
-    __m256 b_floats = _mm256_loadu_ps(b_ptr + i);
+    // Load 8 floats
+    __m256 b_floats = _mm256_loadu_ps(b_ptr + i + 0);
 
     // Conditionally convert based on the type of V::value_type
     __m256i a_ints;
@@ -217,11 +217,13 @@ template <feature_vector V, feature_vector W>
 inline float avx2_inner_product(const V& a, const W& b) {
   // @todo Align on 256 bit boundaries
   const size_t start = 0;
-  const size_t size_a = a.size();
+  const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
-  const auto* a_ptr = a.data();  // Can be either uint8_t* or int8_t*
-  const auto* b_ptr = b.data();  // Can be either uint8_t* or int8_t*
+  // Can be either uint8_t* or int8_t*
+  const auto* a_ptr = a.data();
+  // Can be either uint8_t* or int8_t*
+  const auto* b_ptr = b.data();
 
   __m256 vec_sum = _mm256_setzero_ps();