--- a/src/include/detail/scoring/l2_distance_avx.h
+++ b/src/include/detail/scoring/l2_distance_avx.h
@@ -56,8 +56,8 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
   const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
-  const float* a_ptr = a.data();
-  const float* b_ptr = b.data();
+  const auto* a_ptr = a.data();
+  const auto* b_ptr = b.data();
 
   __m256 vec_sum = _mm256_setzero_ps();
 
@@ -109,15 +109,15 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
  */
 template <feature_vector V, feature_vector W>
   requires std::same_as<typename V::value_type, float> &&
-           std::same_as<typename W::value_type, uint8_t>
+           (std::same_as<typename W::value_type, uint8_t> || std::same_as<typename W::value_type, int8_t>)
 inline float avx2_sum_of_squares(const V& a, const W& b) {
   // @todo Align on 256 bit boundaries
   const size_t start = 0;
   const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
-  const float* a_ptr = a.data();
-  const uint8_t* b_ptr = b.data();
+  const auto* a_ptr = a.data();
+  const auto* b_ptr = b.data();
 
   __m256 vec_sum = _mm256_setzero_ps();
 
@@ -135,7 +135,7 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
     __m256 b_floats = _mm256_cvtepi32_ps(b_ints);
 
     // Subtract floats
-    __m256i diff = _mm256_sub_ps(a_floats, b_floats);
+    __m256 diff = _mm256_sub_ps(a_floats, b_floats);
 
     // Square and add with fmadd
     vec_sum = _mm256_fmadd_ps(diff, diff, vec_sum);
@@ -167,7 +167,7 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
  * uint8_t - float
  */
 template <feature_vector V, feature_vector W>
-  requires std::same_as<typename V::value_type, uint8_t> &&
+  requires (std::same_as<typename V::value_type, uint8_t> || std::same_as<typename V::value_type, int8_t>) &&
            std::same_as<typename W::value_type, float>
 inline float avx2_sum_of_squares(const V& a, const W& b) {
   // @todo Align on 256 bit boundaries
@@ -175,8 +175,8 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
   const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
-  const uint8_t* a_ptr = a.data();
-  const float* b_ptr = b.data();
+  const auto* a_ptr = a.data();
+  const auto* b_ptr = b.data();
 
   __m256 vec_sum = _mm256_setzero_ps();
 
@@ -194,7 +194,7 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
     __m256 a_floats = _mm256_cvtepi32_ps(a_ints);
 
     // Subtract floats
-    __m256i diff = _mm256_sub_ps(a_floats, b_floats);
+    __m256 diff = _mm256_sub_ps(a_floats, b_floats);
 
     // Square and add with fmadd
     vec_sum = _mm256_fmadd_ps(diff, diff, vec_sum);
@@ -226,16 +226,16 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
  * uint8_t - uint8_t
  */
 template <class V, class W>
-  requires std::same_as<typename V::value_type, uint8_t> &&
-           std::same_as<typename W::value_type, uint8_t>
+  requires (std::same_as<typename V::value_type, uint8_t> || std::same_as<typename V::value_type, int8_t>) &&
+           (std::same_as<typename W::value_type, uint8_t> || std::same_as<typename W::value_type, int8_t>)
 inline float avx2_sum_of_squares(const V& a, const W& b) {
   // @todo Align on 256 bit boundaries
   const size_t start = 0;
   const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
-  const uint8_t* a_ptr = a.data();
-  const uint8_t* b_ptr = b.data();
+  const auto* a_ptr = a.data();
+  const auto* b_ptr = b.data();
 
   __m256 vec_sum = _mm256_setzero_ps();
 
@@ -255,10 +255,10 @@ inline float avx2_sum_of_squares(const V& a, const W& b) {
     __m256 b_floats = _mm256_cvtepi32_ps(b_ints);
 
     // Subtract floats
-    __m256i diff = _mm256_sub_ps(a_floats, b_floats);
+    __m256 diff = _mm256_sub_ps(a_floats, b_floats);
 #else
     // Subtract signed integers
-    __m256 i_diff = _mm256_sub_epi32(a_ints, b_ints);
+    __m256i i_diff = _mm256_sub_epi32(a_ints, b_ints);
 
     // Convert integers to floats
     __m256 diff = _mm256_cvtepi32_ps(i_diff);
@@ -309,7 +309,7 @@ inline float avx2_sum_of_squares(const V& a) {
   const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
-  const float* a_ptr = a.data();
+  const auto* a_ptr = a.data();
 
   __m256 vec_sum = _mm256_setzero_ps();
 
@@ -346,14 +346,14 @@ inline float avx2_sum_of_squares(const V& a) {
  * uint8_t
  */
 template <class V>
-  requires std::same_as<typename V::value_type, uint8_t>
+  requires (std::same_as<typename V::value_type, uint8_t> || std::same_as<typename V::value_type, int8_t>)
 inline float avx2_sum_of_squares(const V& a) {
   // @todo Align on 256 bit boundaries
   const size_t start = 0;
   const size_t size_a = size(a);
   const size_t stop = size_a - (size_a % 8);
 
-  const uint8_t* a_ptr = a.data();
+  const auto* a_ptr = a.data();
 
   __m256 vec_sum = _mm256_setzero_ps();