### What
A few changes:
* Updates the build system so that if we have AVX2 support on the host machine, we will build with AVX2 support in the binary. 
* Fixes several bugs in the AVX2 distance functions. 

Based on:
* https://github.com/TileDB-Inc/TileDB-Vector-Search/pull/488
* https://github.com/TileDB-Inc/TileDB-Vector-Search/pull/499

Below are several sets of benchmarks that are run. You can also see more here: https://github.com/TileDB-Inc/TileDB-Vector-Search/pull/499

The summary:
* VAMANA got much faster in both ingestion and query.
* IVF_FLAT had modest improvement.
* IVF_PQ doesn't seem to have a big improvement, and is worse on one benchmark, but not by much.

### CI Benchmarks

C++ CI tests before (left) vs after (right). Note we go from 16 minutes to 10 minutes.

<img width="1505" alt="Screenshot 2024-10-10 at 8 14 13 AM" src="https://github.com/user-attachments/assets/b47bb566-bfec-4560-ba79-b759b838a3ed">

Python CI tests before (left) vs after (right). Note that while the cloud tests are not faster, the non-cloud tests are. i.e. `test/test_distance_metrics.py::test_vamana_l2_distance` goes from 41.01s to 23.19s. And `test/test_ingestion.py::test_ivf_flat_ingestion_f32` goes from 28.90s to 21.48s.

<img width="1363" alt="Screenshot 2024-10-10 at 8 13 13 AM" src="https://github.com/user-attachments/assets/8d04bd0c-c7be-4413-a3e8-7d0172d1b6ae">

### Manual Benchmarks
Before:
![ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/510423fb-1577-46c0-b212-d364ae58f97b)

After:
![ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/82d7ec98-badf-4f0a-97b5-1b5b2e4af6fb)

Before:
![query_time_vs_accuracy](https://github.com/user-attachments/assets/26196978-8a64-40ca-b1a7-c62712975f49)

After:
![query_time_vs_accuracy](https://github.com/user-attachments/assets/5ddd88c7-0d61-4d43-ab2f-2817cd55d996)

### Manual Benchmarks: VAMANA
Before:
![VAMANA_ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/b74e3d83-8dda-433e-97a4-d4aa9e777df6)

After:
![VAMANA_ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/ec77ac97-319a-4d3a-9c76-1c2afb86af26)

Before :
![VAMANA_query_time_vs_accuracy](https://github.com/user-attachments/assets/be50f9e9-0cef-4ba3-be36-bc602130e09c)

After:
![VAMANA_query_time_vs_accuracy](https://github.com/user-attachments/assets/b79c07ac-6a8c-450c-915b-d194ebc44626)

### Manual Benchmarks: IVF_FLAT
Before:
![IVF_FLAT_ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/50152bfb-cc76-48f7-99b5-e3860b5ec260)

After:
![IVF_FLAT_ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/e7689ba5-2848-401f-91aa-9f8b10190c89)

Before:
![IVF_FLAT_query_time_vs_accuracy](https://github.com/user-attachments/assets/6dec4e55-9a5d-48e8-bdf1-1674119f08be)

After:
![IVF_FLAT_query_time_vs_accuracy](https://github.com/user-attachments/assets/92d8bfcb-9a16-4e7c-8e7a-5b080da5381c)

### Manual Benchmarks: IVF_PQ
Before:
![IVF_PQ_ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/f942eec9-8317-4e9d-9d22-d2e12b537c85)

After:
![IVF_PQ_ingestion_time_vs_accuracy](https://github.com/user-attachments/assets/b8ba2b20-ca10-466b-996e-eed385f1c6ce)

Before :
![IVF_PQ_query_time_vs_accuracy](https://github.com/user-attachments/assets/7fa5e18f-d9f0-4d91-997c-9b1b9bcaee01)

After:
![IVF_PQ_query_time_vs_accuracy](https://github.com/user-attachments/assets/6d44c77a-50ca-4186-8d7c-521785b61a18)